fix: Voice over focus on add favorite flow
