+++ src/place-screen/components/EstimatedCallList.tsx
@@ -2,7 +2,7 @@ import {GenericSectionItem, SectionSeparator} from '@atb/components/sections';
 import {EstimatedCall} from '@atb/api/types/departures';
 import {ThemeText} from '@atb/components/text';
 import {FlatList} from 'react-native-gesture-handler';
-import React, {RefObject, useCallback, useRef} from 'react';
+import React, {RefObject, useCallback} from 'react';
 import {DeparturesTexts, useTranslation} from '@atb/translations';
 import {
   EstimatedCallItem,
@@ -48,13 +48,12 @@ export const EstimatedCallList = ({
     addedFavoritesVisibleOnDashboard,
   );
 
-  // TODO: This ref isn't hooked up anywhere for now. Since the favorite feature
-  // is due to be moved out of EstimatedCallList, it should be fixed when that
-  // happens. (https://github.com/AtB-AS/kundevendt/issues/19124)
-  const onCloseFocusRef = useRef<RefObject<any>>(null);
-
   const onPressFavorite = useCallback(
-    (departure: EstimatedCall, existingFavorite?: StoredFavoriteDeparture) =>
+    (
+      departure: EstimatedCall,
+      existingFavorite: StoredFavoriteDeparture | undefined,
+      onCloseRef: RefObject<any>,
+    ) =>
       onMarkFavourite(
         {
           ...departure.serviceJourney.line,
@@ -62,7 +61,7 @@ export const EstimatedCallList = ({
           destinationDisplay: departure.destinationDisplay,
         },
         existingFavorite,
-        onCloseFocusRef,
+        onCloseRef,
       ),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [],
